re-arrange folder format
