+++ components/navigation/Navbar.tsx
@@ -1,12 +1,90 @@
+/* eslint-disable react/require-default-props */
 /* eslint-disable import/no-unresolved */
-import { useState } from 'react';
+import { Dispatch, SetStateAction, useState } from 'react';
 import { useTheme } from 'next-themes';
 import Image from 'next/image';
 import Link from 'next/link';
+import router from 'next/router';
 
 import images from '../../assets';
-import ButtonGroup from './ButtonGroup';
-import MenuItems from './MenuItems';
+import Button from '../Button';
+
+// Button Group
+interface IButtonGroup {
+  setActive: Dispatch<SetStateAction<string>>;
+}
+
+const ButtonGroup = ({ setActive }: IButtonGroup) => {
+  const hasConnected = true;
+  return hasConnected ? (
+    <Button
+      btnName="Create"
+      classStyles="mx-2 rounded-xl"
+      handleClick={() => {
+        setActive('');
+        router.push('/create-nft');
+      }}
+    />
+  ) : (
+    <Button
+      btnName="Connect"
+      classStyles="mx-2 rounded-xl"
+      handleClick={() => {}}
+    />
+  );
+};
+
+// Menu Items
+interface IMenuItemsProps {
+  isMobile?: boolean;
+  active: string;
+  setActive: Dispatch<SetStateAction<string>>;
+}
+
+const MenuItems = ({
+  isMobile = false,
+  active,
+  setActive,
+}: IMenuItemsProps) => {
+  const generateLink = (i: number) => {
+    switch (i) {
+      case 0:
+        return '/';
+      case 1:
+        return '/create-goods';
+      case 2:
+        return '/my-goods';
+      default:
+        return '/';
+    }
+  };
+
+  return (
+    <ul
+      className={`list-none flexCenter flex-row ${
+        isMobile && 'flex-col h-full'
+      }`}
+    >
+      {['Explore Goods', 'Listed Goods', 'My Goods'].map((item, i) => (
+        <li
+          key={i}
+          onClick={() => {
+            setActive(item);
+          }}
+          className={`flex flex-row items-center font-poppins font-semibold text-base dark:hover:text-white hover:text-nft-dark mx-3
+            ${
+              active === item
+                ? 'dark:text-white text-nft-black-1'
+                : 'dark:text-nft-gray-3 text-nft-gray-2'
+            } 
+            ${isMobile && 'my-5 text-xl'}`}
+        >
+          <Link href={generateLink(i)}>{item}</Link>
+        </li>
+      ))}
+    </ul>
+  );
+};
 
 const Navbar = () => {
   const { theme, setTheme } = useTheme();
