+++ components/views/create-goods/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const CreateGoods = () => (
+  <div>CreateGoods</div>
+);
+
+export default CreateGoods;
