+++ pages/_app.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable import/no-unresolved */
 import Script from 'next/script';
 import { ThemeProvider } from 'next-themes';
-import { Navbar, Footer } from '../components';
 import '../styles/globals.css';
+import RgmsLayout from '../components/layout/RgmsLayout';
 
 const MyApp = ({
   Component,
@@ -13,9 +13,9 @@ const MyApp = ({
 }) => (
   <ThemeProvider attribute="class">
     <div className="dark:bg-nft-dark bg-white min-h-screen">
-      <Navbar />
-      <Component {...pageProps} />
-      <Footer />
+      <RgmsLayout>
+        <Component {...pageProps} />
+      </RgmsLayout>
     </div>
     <Script
       src="https://kit.fontawesome.com/04aa5e1825.js"
