--- pages/created-goods.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const CreatedGoods = () => (
-  <div>CreatedGoods</div>
-);
-
-export default CreatedGoods;
